Add an :after method to setf the node-value for element objects
It is currently not defined what the (setf (dom:node-value) ...)
method should do for objects of type element, and thus, in the
xpath library, the method does nothing.

This commit creates an :after method that simply removes all
children of the element and replaces them with a single object of
type text with the value to be set.